chore: update folder structure of footer for hk and row
update ROW footer navigation contents, description, and refactor the folder structure of HK, ROW footer
